+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # DocumenterTools.jl changelog
 
+## Unreleased
+
+* DocumenterTools now requires at least Julia 1.6.
+* ![Enhancement][badge-enhancement] `DocumenterTools.genkeys` function now uses the `OpenSSH_jll` for the `ssh-keygen` binary and therefore no longer requires the user to have OpenSSH installed. ([#36][github-36], [#56][github-56])
+
 ## Version `v0.1.13`
 
 * ![Enhancement][badge-enhancement] The wording of the text in the the old version warning box was improved. ([#54][github-54])
@@ -66,6 +71,7 @@ Maintenance release declaring compatibility with Documenter 0.25. ([#39][github-
 [github-32]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/32
 [github-33]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/33
 [github-35]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/35
+[github-36]: https://github.com/JuliaDocs/DocumenterTools.jl/issues/36
 [github-39]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/39
 [github-40]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/40
 [github-43]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/43
@@ -75,6 +81,7 @@ Maintenance release declaring compatibility with Documenter 0.25. ([#39][github-
 [github-51]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/51
 [github-52]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/52
 [github-54]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/54
+[github-56]: https://github.com/JuliaDocs/DocumenterTools.jl/pull/56
 
 
 [badge-breaking]: https://img.shields.io/badge/BREAKING-red.svg